added a new function to manager interface

+++ src/model/Menu.java
@@ -7,15 +7,15 @@ import java.util.Map;
 
 public class Menu {
 
-	private Map<MenuItem, Double> items;
+	private Map<MenuItem, Long> items;
 	private List<PizzaTopping> pizzaToppings;
 	
 	public Menu(){
-		this.items = new HashMap<MenuItem, Double>();
+		this.items = new HashMap<MenuItem, Long>();
 		this.pizzaToppings = new ArrayList<PizzaTopping>();
 	}
 	
-	public Map<MenuItem, Double> getMenuItems(){
+	public Map<MenuItem, Long> getMenuItems(){
 		return this.items;
 	}
